Module libax::cbindings

Available on crate feature cbindings only.

Exported C bindings, to call ArceOS funtions from C code.

Functions

ax_accept⚠net

Accept for connections on a socket

ax_bind⚠net

Bind a address to a socket.

ax_clock_gettime⚠

Get clock time since booting

ax_close⚠alloc

Close a file by fd.

ax_connect⚠net

Connects the socket to the address specified.

ax_dup⚠alloc

Duplicate a file descriptor

ax_dup3⚠alloc

dup3() is the same as dup2(), except that:

ax_errno_string⚠

Returns a pointer to the string representation of the given error code.

ax_exit⚠

Exits the current thread.

ax_fcntl⚠alloc

Fcntl implementation

ax_free⚠alloc

Deallocate memory.

ax_fstat⚠alloc

Get file metadata by fd and write into buf.

ax_getcwd⚠fs

Get the path of the current directory.

ax_getpid⚠multitask

Get current thread ID

ax_listen⚠net

Listen for connections on a socket

ax_lseek⚠fs

Set the position of the file indicated by fd.

ax_lstat⚠fs

Get the metadata of the symbolic link and write into buf.

ax_malloc⚠alloc

Allocate memory and return the memory address.

ax_nanosleep⚠

Sleep some nanoseconds

ax_open⚠fs

Open a file by filename and insert it into the file descriptor table.

ax_panic⚠

Abort the current process.

ax_pipe⚠pipe

Create a pipe

ax_print_str⚠

Print a string to the global standard output stream.

ax_println_str⚠

Printf a string to the global standard output stream. Add a line break.

ax_rand_u32⚠

Returns a 32-bit unsigned pseudo random interger.

ax_read⚠alloc

Read data from the file indicated by fd.

ax_recv⚠net

Receive a message on a socket.

ax_recvfrom⚠net

Receive a message on a socket and get its source address.

ax_resolve_sockaddr⚠net

Query addresses for a domain name.

ax_send⚠net

Send a message on a socket to the address connected.

ax_sendto⚠net

Send a message on a socket to the address specified.

ax_shutdown⚠net

Shut down a full-duplex connection.

ax_socket⚠net

Create an socket for communication.

ax_srand⚠

Sets the seed for the random number generator.

ax_stat⚠fs

Get the file metadata by path and write into buf.

ax_sysconf⚠

Return system configuration infomation

ax_write⚠alloc

Write data to the file indicated by fd.