Struct libax::fs::File

pub struct File { /* private fields */ }

Available on crate feature fs only.

An object providing access to an open file on the filesystem.

Implementations

impl File

pub fn open(path: &str) -> Result<File, AxError>

Attempts to open a file in read-only mode.

pub fn create(path: &str) -> Result<File, AxError>

Opens a file in write-only mode.

pub fn create_new(path: &str) -> Result<File, AxError>

Creates a new file in read-write mode; error if the file exists.

pub fn options() -> OpenOptions

Returns a new OpenOptions object.

pub fn set_len(&self, size: u64) -> Result<(), AxError>

Truncates or extends the underlying file, updating the size of this file to become size.

pub fn metadata(&self) -> Result<Metadata, AxError>

Queries metadata about the underlying file.