Struct libax::io::BufReader

pub struct BufReader<R> { /* private fields */ }

The BufReader<R> struct adds buffering to any reader.

Implementations

impl<R> BufReader<R>where R: Read,

pub const fn new(inner: R) -> BufReader<R>

Creates a new BufReader<R> with a default buffer capacity (1 KB).

impl<R> BufReader<R>

pub const fn get_ref(&self) -> &R

Gets a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Gets a mutable reference to the underlying reader.

pub fn buffer(&self) -> &[u8]

Returns a reference to the internally buffered data.

Unlike fill_buf, this will not attempt to fill the buffer if it is empty.

pub const fn capacity(&self) -> usize

Returns the number of bytes the internal buffer can hold at once.

pub fn into_inner(self) -> R

Unwraps this BufReader<R>, returning the underlying reader.