Trait libax::net::ToSocketAddrs

pub trait ToSocketAddrs {
    type Iter: Iterator<Item = SocketAddr>;

    // Required method
    fn to_socket_addrs(&self) -> Result<Self::Iter>;
}

Available on crate feature net only.

A trait for objects which can be converted or resolved to one or more SocketAddr values.

Required Associated Types

type Iter: Iterator<Item = SocketAddr>

Returned iterator over socket addresses which this type may correspond to.

Required Methods

fn to_socket_addrs(&self) -> Result<Self::Iter>

Converts this object to an iterator of resolved SocketAddrs.